refactor: Move useImageLoad into own directory
