+++ packages/insight-viewer/src/hooks/useImageLoad/const.ts
@@ -1,0 +1,6 @@
+export const LOADING_STATUS = {
+  INITIAL: 'initial',
+  LOADING: 'loading',
+  SUCCESS: 'success',
+  FAIL: 'fail',
+} as const
