+++ packages/insight-viewer/src/hooks/useImageLoad/types.ts
@@ -1,0 +1,3 @@
+import { LOADING_STATUS } from './const'
+
+export type LoadingStatus = typeof LOADING_STATUS[keyof typeof LOADING_STATUS]
